+++ Box.V2.Samples.JWTAuth/Program.cs
@@ -33,7 +33,8 @@ namespace Box.V2.Samples.JWTAuth
             // rename the private_key.pem.example to private_key.pem and put your JWT private key in the file
             var privateKey = File.ReadAllText("private_key.pem.example");
 
-            var boxConfig = new BoxConfig(CLIENT_ID, CLIENT_SECRET, ENTERPRISE_ID, privateKey, JWT_PRIVATE_KEY_PASSWORD, JWT_PUBLIC_KEY_ID);
+            var boxConfig = new BoxConfigBuilder(CLIENT_ID, CLIENT_SECRET, ENTERPRISE_ID, privateKey, JWT_PRIVATE_KEY_PASSWORD, JWT_PUBLIC_KEY_ID)
+                .Build();
             var boxJWT = new BoxJWTAuth(boxConfig);
 
             var adminToken = boxJWT.AdminToken();