+++ Box.V2/Config/BoxConfig.cs
@@ -7,9 +7,9 @@ using System.Text;
 
 namespace Box.V2.Config
 {
-    public class BoxConfig : IBoxConfig
+    public sealed class BoxConfig : IBoxConfig
     {
-        private static string DefaultUserAgent = "Box Windows SDK v" + AssemblyInfo.NuGetVersion;
+        internal static string DefaultUserAgent = "Box Windows SDK v" + AssemblyInfo.NuGetVersion;
 
         /// <summary>
         /// Instantiates a Box config with all of the standard defaults
@@ -44,6 +44,26 @@ namespace Box.V2.Config
             JWTPrivateKeyPassword = jwtPrivateKeyPassword;
             JWTPublicKeyId = jwtPublicKeyId;
             UserAgent = DefaultUserAgent;
+        }
+
+        public BoxConfig(BoxConfigBuilder builder)
+        {
+            ClientId = builder.ClientId;
+            ClientSecret = builder.ClientSecret;
+            EnterpriseId = builder.EnterpriseId;
+            JWTPrivateKey = builder.JWTPrivateKey;
+            JWTPrivateKeyPassword = builder.JWTPrivateKeyPassword;
+            JWTPublicKeyId = builder.JWTPublicKeyId;
+            UserAgent = builder.UserAgent;
+            BoxApiHostUri = builder.BoxApiHostUri;
+            BoxAccountApiHostUri = builder.BoxAccountApiHostUri;
+            BoxApiUri = builder.BoxApiUri;
+            BoxUploadApiUri = builder.BoxUploadApiUri;
+            RedirectUri = builder.RedirectUri;
+            DeviceId = builder.DeviceId;
+            DeviceName = builder.DeviceName;
+            AcceptEncoding = builder.AcceptEncoding;
+            WebProxy = builder.WebProxy;
         }
 
         /// <summary>
@@ -122,67 +142,67 @@ namespace Box.V2.Config
             return new BoxConfig(clientId, clientSecret, enterpriseId, privateKey, rsaSecret, publicKeyId);
         }
 
-        public virtual Uri BoxApiHostUri { get; set; } = new Uri(Constants.BoxApiHostUriString);
-        public virtual Uri BoxAccountApiHostUri { get; set; } = new Uri(Constants.BoxAccountApiHostUriString);
-        public virtual Uri BoxApiUri { get; set; } = new Uri(Constants.BoxApiUriString);
-        public virtual Uri BoxUploadApiUri { get; set; } = new Uri(Constants.BoxUploadApiUriString);
+        public Uri BoxApiHostUri { get; private set; } = new Uri(Constants.BoxApiHostUriString);
+        public Uri BoxAccountApiHostUri { get; private set; } = new Uri(Constants.BoxAccountApiHostUriString);
+        public Uri BoxApiUri { get; private set; } = new Uri(Constants.BoxApiUriString);
+        public Uri BoxUploadApiUri { get; private set; } = new Uri(Constants.BoxUploadApiUriString);
 
-        public virtual string ClientId { get; private set; }
-        public virtual string ConsumerKey { get; private set; }
-        public virtual string ClientSecret { get; private set; }
-        public virtual Uri RedirectUri { get; set; }
+        public string ClientId { get; private set; }
+        public string ConsumerKey { get; private set; }
+        public string ClientSecret { get; private set; }
+        public Uri RedirectUri { get; private set; }
 
         public string EnterpriseId { get; private set; }
         public string JWTPrivateKey { get; private set; }
         public string JWTPrivateKeyPassword { get; private set; }
         public string JWTPublicKeyId { get; private set; }
 
-        public string DeviceId { get; set; }
-        public string DeviceName { get; set; }
-        public string UserAgent { get; set; }
+        public string DeviceId { get; private set; }
+        public string DeviceName { get; private set; }
+        public string UserAgent { get; private set; }
 
         /// <summary>
         /// Sends compressed responses from Box for faster response times
         /// </summary>
-        public CompressionType? AcceptEncoding { get; set; }
+        public CompressionType? AcceptEncoding { get; private set; }
 
-        public virtual Uri AuthCodeBaseUri { get { return new Uri(BoxAccountApiHostUri, Constants.AuthCodeString); } }
-        public virtual Uri AuthCodeUri { get { return new Uri(AuthCodeBaseUri, string.Format("?response_type=code&client_id={0}&redirect_uri={1}", ClientId, RedirectUri)); } }
-        public virtual Uri FoldersEndpointUri { get { return new Uri(BoxApiUri, Constants.FoldersString); } }
-        public virtual Uri TermsOfServicesUri { get { return new Uri(BoxApiUri, Constants.TermsOfServicesString); } }
-        public virtual Uri TermsOfServiceUserStatusesUri { get { return new Uri(BoxApiUri, Constants.TermsOfServiceUserStatusesString); } }
-        public virtual Uri FilesEndpointUri { get { return new Uri(BoxApiUri, Constants.FilesString); } }
-        public virtual Uri FilesUploadEndpointUri { get { return new Uri(BoxUploadApiUri, Constants.FilesUploadString); } }
+        public Uri AuthCodeBaseUri { get { return new Uri(BoxAccountApiHostUri, Constants.AuthCodeString); } }
+        public Uri AuthCodeUri { get { return new Uri(AuthCodeBaseUri, string.Format("?response_type=code&client_id={0}&redirect_uri={1}", ClientId, RedirectUri)); } }
+        public Uri FoldersEndpointUri { get { return new Uri(BoxApiUri, Constants.FoldersString); } }
+        public Uri TermsOfServicesUri { get { return new Uri(BoxApiUri, Constants.TermsOfServicesString); } }
+        public Uri TermsOfServiceUserStatusesUri { get { return new Uri(BoxApiUri, Constants.TermsOfServiceUserStatusesString); } }
+        public Uri FilesEndpointUri { get { return new Uri(BoxApiUri, Constants.FilesString); } }
+        public Uri FilesUploadEndpointUri { get { return new Uri(BoxUploadApiUri, Constants.FilesUploadString); } }
 
         /// <summary>
         /// Upload session
         /// </summary>
-        public virtual Uri FilesUploadSessionEndpointUri { get { return new Uri(BoxUploadApiUri, Constants.FilesUploadSessionString); } }
-        public virtual Uri FilesPreflightCheckUri { get { return new Uri(BoxApiUri, Constants.FilesUploadString); } }
-        public virtual Uri CommentsEndpointUri { get { return new Uri(BoxApiUri, Constants.CommentsString); } }
-        public virtual Uri SearchEndpointUri { get { return new Uri(BoxApiUri, Constants.SearchString); } }
-        public virtual Uri UserEndpointUri { get { return new Uri(BoxApiUri, Constants.UserString); } }
-        public virtual Uri InviteEndpointUri { get { return new Uri(BoxApiUri, Constants.InviteString); } }
-        public virtual Uri CollaborationsEndpointUri { get { return new Uri(BoxApiUri, Constants.CollaborationsString); } }
-        public virtual Uri GroupsEndpointUri { get { return new Uri(BoxApiUri, Constants.GroupsString); } }
-        public virtual Uri GroupMembershipEndpointUri { get { return new Uri(BoxApiUri, Constants.GroupMembershipString); } }
-        public virtual Uri RetentionPoliciesEndpointUri { get { return new Uri(BoxApiUri, Constants.RetentionPoliciesString); } }
-        public virtual Uri RetentionPolicyAssignmentsUri { get { return new Uri(BoxApiUri, Constants.RetentionPolicyAssignmentsString); } }
-        public virtual Uri FileVersionRetentionsUri { get { return new Uri(BoxApiUri, Constants.FileVersionRetentionsString); } }
-        public virtual Uri EventsUri { get { return new Uri(BoxApiUri, Constants.EventsString); } }
-        public virtual Uri MetadataTemplatesUri { get { return new Uri(BoxApiUri, Constants.MetadataTemplatesString); } }
-        public virtual Uri CreateMetadataTemplateUri { get { return new Uri(BoxApiUri, Constants.CreateMetadataTemplateString); } }
-        public virtual Uri MetadataQueryUri { get { return new Uri(BoxApiUri, Constants.MetadataQueryString); } }
-        public virtual Uri WebhooksUri { get { return new Uri(BoxApiUri, Constants.WebhooksString); } }
-        public virtual Uri RecentItemsUri { get { return new Uri(BoxApiUri, Constants.RecentItemsString); } }
-        public virtual Uri EnterprisesUri { get { return new Uri(BoxApiUri, Constants.EnterprisesString); } }
-        public virtual Uri DevicePinUri { get { return new Uri(BoxApiUri, Constants.DevicePinString); } }
-        public virtual Uri CollaborationWhitelistEntryUri { get { return new Uri(BoxApiUri, Constants.CollaborationWhitelistEntryString); } }
-        public virtual Uri CollaborationWhitelistTargetEntryUri { get { return new Uri(BoxApiUri, Constants.CollaborationWhitelistTargetEntryString); } }
-        public virtual Uri MetadataCascadePolicyUri { get { return new Uri(BoxApiUri, Constants.MetadataCascadePoliciesString); } }
-        public virtual Uri StoragePoliciesUri { get { return new Uri(BoxApiUri, Constants.StoragePoliciesString); } }
-        public virtual Uri StoragePolicyAssignmentsUri { get { return new Uri(BoxApiUri, Constants.StoragePolicyAssignmentsString); } }
-        public virtual Uri StoragePolicyAssignmentsForTargetUri { get { return new Uri(BoxApiUri, Constants.StoragePolicyAssignmentsForTargetString); } }
+        public Uri FilesUploadSessionEndpointUri { get { return new Uri(BoxUploadApiUri, Constants.FilesUploadSessionString); } }
+        public Uri FilesPreflightCheckUri { get { return new Uri(BoxApiUri, Constants.FilesUploadString); } }
+        public Uri CommentsEndpointUri { get { return new Uri(BoxApiUri, Constants.CommentsString); } }
+        public Uri SearchEndpointUri { get { return new Uri(BoxApiUri, Constants.SearchString); } }
+        public Uri UserEndpointUri { get { return new Uri(BoxApiUri, Constants.UserString); } }
+        public Uri InviteEndpointUri { get { return new Uri(BoxApiUri, Constants.InviteString); } }
+        public Uri CollaborationsEndpointUri { get { return new Uri(BoxApiUri, Constants.CollaborationsString); } }
+        public Uri GroupsEndpointUri { get { return new Uri(BoxApiUri, Constants.GroupsString); } }
+        public Uri GroupMembershipEndpointUri { get { return new Uri(BoxApiUri, Constants.GroupMembershipString); } }
+        public Uri RetentionPoliciesEndpointUri { get { return new Uri(BoxApiUri, Constants.RetentionPoliciesString); } }
+        public Uri RetentionPolicyAssignmentsUri { get { return new Uri(BoxApiUri, Constants.RetentionPolicyAssignmentsString); } }
+        public Uri FileVersionRetentionsUri { get { return new Uri(BoxApiUri, Constants.FileVersionRetentionsString); } }
+        public Uri EventsUri { get { return new Uri(BoxApiUri, Constants.EventsString); } }
+        public Uri MetadataTemplatesUri { get { return new Uri(BoxApiUri, Constants.MetadataTemplatesString); } }
+        public Uri CreateMetadataTemplateUri { get { return new Uri(BoxApiUri, Constants.CreateMetadataTemplateString); } }
+        public Uri MetadataQueryUri { get { return new Uri(BoxApiUri, Constants.MetadataQueryString); } }
+        public Uri WebhooksUri { get { return new Uri(BoxApiUri, Constants.WebhooksString); } }
+        public Uri RecentItemsUri { get { return new Uri(BoxApiUri, Constants.RecentItemsString); } }
+        public Uri EnterprisesUri { get { return new Uri(BoxApiUri, Constants.EnterprisesString); } }
+        public Uri DevicePinUri { get { return new Uri(BoxApiUri, Constants.DevicePinString); } }
+        public Uri CollaborationWhitelistEntryUri { get { return new Uri(BoxApiUri, Constants.CollaborationWhitelistEntryString); } }
+        public Uri CollaborationWhitelistTargetEntryUri { get { return new Uri(BoxApiUri, Constants.CollaborationWhitelistTargetEntryString); } }
+        public Uri MetadataCascadePolicyUri { get { return new Uri(BoxApiUri, Constants.MetadataCascadePoliciesString); } }
+        public Uri StoragePoliciesUri { get { return new Uri(BoxApiUri, Constants.StoragePoliciesString); } }
+        public Uri StoragePolicyAssignmentsUri { get { return new Uri(BoxApiUri, Constants.StoragePolicyAssignmentsString); } }
+        public Uri StoragePolicyAssignmentsForTargetUri { get { return new Uri(BoxApiUri, Constants.StoragePolicyAssignmentsForTargetString); } }
 
         /// <summary>
         /// Gets the shared items endpoint URI.
@@ -190,7 +210,7 @@ namespace Box.V2.Config
         /// <value>
         /// The shared items endpoint URI.
         /// </value>
-        public virtual Uri SharedItemsUri { get { return new Uri(BoxApiUri, Constants.SharedItemsString); } }
+        public Uri SharedItemsUri { get { return new Uri(BoxApiUri, Constants.SharedItemsString); } }
 
         /// <summary>
         /// Gets the task assignments endpoint URI.
@@ -198,12 +218,12 @@ namespace Box.V2.Config
         /// <value>
         /// The task assignments endpoint URI.
         /// </value>
-        public virtual Uri TaskAssignmentsEndpointUri { get { return new Uri(BoxApiUri, Constants.TaskAssignmentsString); } }
+        public Uri TaskAssignmentsEndpointUri { get { return new Uri(BoxApiUri, Constants.TaskAssignmentsString); } }
 
         /// <summary>
         /// Gets the tasks endpoint URI.
         /// </summary>
-        public virtual Uri TasksEndpointUri { get { return new Uri(BoxApiUri, Constants.TasksString); } }
+        public Uri TasksEndpointUri { get { return new Uri(BoxApiUri, Constants.TasksString); } }
 
         /// <summary>
         /// Gets the collections endpoint URI.
@@ -211,35 +231,35 @@ namespace Box.V2.Config
         /// <value>
         /// The collections endpoint URI.
         /// </value>
-        public virtual Uri CollectionsEndpointUri { get { return new Uri(BoxApiUri, Constants.CollectionsString); } }
+        public Uri CollectionsEndpointUri { get { return new Uri(BoxApiUri, Constants.CollectionsString); } }
         /// <summary>
         /// Gets the web links endpoint URI.
         /// </summary>
-        public virtual Uri WebLinksEndpointUri { get { return new Uri(BoxApiUri, Constants.WebLinksString); } }
+        public Uri WebLinksEndpointUri { get { return new Uri(BoxApiUri, Constants.WebLinksString); } }
         /// <summary>
         /// Gets the legal hold policies endpoint URI.
         /// </summary>
-        public virtual Uri LegalHoldPoliciesEndpointUri { get { return new Uri(BoxApiUri, Constants.LegalHoldPoliciesString); } }
+        public Uri LegalHoldPoliciesEndpointUri { get { return new Uri(BoxApiUri, Constants.LegalHoldPoliciesString); } }
         /// <summary>
         /// Gets the legal hold policies endpoint URI.
         /// </summary>
-        public virtual Uri LegalHoldPolicyAssignmentsEndpointUri { get { return new Uri(BoxApiUri, Constants.LegalHoldPolicyAssignmentsString); } }
+        public Uri LegalHoldPolicyAssignmentsEndpointUri { get { return new Uri(BoxApiUri, Constants.LegalHoldPolicyAssignmentsString); } }
         /// <summary>
         /// Gets the file viersion legal holds endpoint URI.
         /// </summary>
-        public virtual Uri FileVersionLegalHoldsEndpointUri { get { return new Uri(BoxApiUri, Constants.FileVersionLegalHoldsString); } }
+        public Uri FileVersionLegalHoldsEndpointUri { get { return new Uri(BoxApiUri, Constants.FileVersionLegalHoldsString); } }
         /// <summary>
         /// Gets the zip downloads endpoint URI.
         /// </summary>
-        public virtual Uri ZipDownloadsEndpointUri { get { return new Uri(BoxApiUri, Constants.ZipDownloadsString); } }
+        public Uri ZipDownloadsEndpointUri { get { return new Uri(BoxApiUri, Constants.ZipDownloadsString); } }
         /// <summary>
         /// Gets the folder locks endpoint URI.
         /// </summary>
-        public virtual Uri FolderLocksEndpointUri { get { return new Uri(BoxApiUri, Constants.FolderLocksString); } }
+        public Uri FolderLocksEndpointUri { get { return new Uri(BoxApiUri, Constants.FolderLocksString); } }
         /// <summary>
         /// The web proxy for HttpRequestHandler
         /// </summary>
-        public IWebProxy WebProxy { get; set; }
+        public IWebProxy WebProxy { get; private set; }
     }
 
     public enum CompressionType